+++ app/src/main/java/com/example/timer/TimeFormat.kt
@@ -1,0 +1,10 @@
+package com.example.timer
+
+fun addZero(digit: Long): String{
+    return if (digit in 0..9) {
+        "0$digit"
+    } else {
+        "$digit"
+    }
+}
+
