+++ app/src/main/java/com/example/timer/ui/theme/ButtonState.kt
@@ -1,0 +1,7 @@
+package com.example.timer.ui.theme
+
+enum class ButtonState {
+    Play,
+    Pause,
+    Restart
+}
