+++ app/src/main/java/com/example/timer/ui/theme/Screen.kt
@@ -1,0 +1,6 @@
+package com.example.timer.ui.theme
+
+enum class Screen {
+    Timer,
+    Input
+}
